Refactor spritesheet and pixel renderers into packages
