+++ docs/research/reactive_switch_state.md
@@ -24,7 +24,7 @@ contacting the manager on every frame.
   hooks to `on_switch_state`.
 - Display renderers such as
   `src/heart/display/renderers/text.py`, `life.py`,
-  `spritesheet.py`, `spritesheet_random.py`, `multi_scene.py`, and
+  `spritesheet/renderer.py`, `spritesheet_random/renderer.py`, `multi_scene.py`, and
   `yolisten.py` enable the cache and consume `get_switch_state()` during
   their render loops.
 - Navigation controllers in `src/heart/navigation.py` call
