+++ src/heart/display/renderers/pixels/__init__.py
@@ -1,0 +1,3 @@
+from heart.display.renderers.pixels.renderer import Border as Border
+from heart.display.renderers.pixels.renderer import Rain as Rain
+from heart.display.renderers.pixels.renderer import Slinky as Slinky
