+++ src/heart/display/renderers/sliding_image/state.py
@@ -2,12 +2,15 @@ from __future__ import annotations
 
 from dataclasses import dataclass
 
+import pygame
+
 
 @dataclass(frozen=True)
 class SlidingImageState:
     offset: int = 0
     speed: int = 1
     width: int = 0
+    image: pygame.Surface | None = None
 
     def advance(self) -> "SlidingImageState":
         if self.width <= 0:
