+++ src/heart/display/renderers/spritesheet/__init__.py
@@ -1,0 +1,10 @@
+from heart.display.renderers.spritesheet.renderer import \
+    SpritesheetLoop as SpritesheetLoop
+from heart.display.renderers.spritesheet.renderer import \
+    create_spritesheet_loop as create_spritesheet_loop
+from heart.display.renderers.spritesheet.state import \
+    BoundingBox as BoundingBox
+from heart.display.renderers.spritesheet.state import \
+    FrameDescription as FrameDescription
+from heart.display.renderers.spritesheet.state import LoopPhase as LoopPhase
+from heart.display.renderers.spritesheet.state import Size as Size
