+++ src/heart/display/renderers/spritesheet_random/__init__.py
@@ -1,0 +1,4 @@
+from heart.display.renderers.spritesheet_random.renderer import \
+    SpritesheetLoopRandom as SpritesheetLoopRandom
+from heart.display.renderers.spritesheet_random.renderer import \
+    create_spritesheet_loop_random as create_spritesheet_loop_random
